pyproject.toml
[tool.poetry]
name = "picasso-ii"
version = "0.1.0"
description = "test work"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.1"
djangorestframework = "3.14.00"
psycopg2-binary = "2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"


[tool.poetry.group.test.dependencies]
flake8 = "^7.1.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
